% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbisant.R
\docType{data}
\name{orbisant_colours}
\alias{orbisant_colours}
\title{General colour palette imports as hex codes}
\format{
An object of class \code{character} of length 6.
}
\usage{
orbisant_colours
}
\description{
General colour palette imports as hex codes
}
\keyword{datasets}
