% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbisant.R
\docType{data}
\name{orbisant_palette}
\alias{orbisant_palette}
\title{General colour palette imports as hex codes
Uses the colours outlined here plus some others that are graphically related https://www.canva.com/colors/color-palettes/floaty-pool-fun/}
\format{
An object of class \code{character} of length 6.
}
\usage{
orbisant_palette
}
\description{
General colour palette imports as hex codes
Uses the colours outlined here plus some others that are graphically related https://www.canva.com/colors/color-palettes/floaty-pool-fun/
}
\keyword{datasets}
