% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbisant.R
\name{scale_colour_orbisant}
\alias{scale_colour_orbisant}
\alias{scale_color_orbisant}
\title{Orbisant colour scale}
\usage{
scale_colour_orbisant(discrete = TRUE, reverse = FALSE, ...)

scale_color_orbisant(discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{discrete}{Whether the colour aesthetic is discrete or not}

\item{reverse}{Whether the colour palette should be reversed}

\item{...}{Additional arguments}
}
\description{
Orbisant colour scale
}
\examples{
library(ggplot2)
ggplot(university_enrolments, aes(x = domestic, y = international, colour = state)) +
 geom_point() +
 scale_colour_orbisant() +
 theme_orbisant()
 
}
