% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbisant.R
\name{theme_orbisant}
\alias{theme_orbisant}
\title{Styles plots according to Orbisant's theme}
\usage{
theme_orbisant()
}
\description{
Uses the colours outlined here plus some 2 others https://www.canva.com/colors/color-palettes/floaty-pool-fun/
}
\examples{
library(ggplot2)
ggplot(university_enrolments, aes(x = domestic, y = international)) +
 geom_point() +
 theme_orbisant()
}
