% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbisant.R
\name{theme_orbisant}
\alias{theme_orbisant}
\title{Style general plot features according to the Orbisant theme}
\usage{
theme_orbisant()
}
\description{
Uses the colours outlined here plus some 2 others https://www.canva.com/colors/color-palettes/floaty-pool-fun/
}
